+++ coolamqp/attaches/declarer.py
@@ -297,7 +297,7 @@ class Declarer(Channeler, Synchronized):
 
         if persistent:
             if obj not in self.declared:
-                self.declared.add(obj)  # todo access not threadsafe
+                self.declared.add(obj)
 
         self.left_to_declare.append(Operation(self, obj, fut, span, enqueued_span))
         self._do_operations()
@@ -311,8 +311,7 @@ class Declarer(Channeler, Synchronized):
 
         To be called when it's possible that something can be done
         """
-        if (self.state != ST_ONLINE) or len(self.left_to_declare) == 0 or (
-                self.in_process is not None):
+        if self.state != ST_ONLINE or not len(self.left_to_declare) or self.in_process is not None:
             return
 
         self.in_process = self.left_to_declare.popleft()