+++ stress_tests/settings.py
@@ -6,7 +6,7 @@ from coolamqp.objects import NodeDefinition
 
 logger = logging.getLogger(__name__)
 
-NODE = NodeDefinition(os.environ.get('AMQP_HOST', '127.0.0.1'), 'guest', 'guest', heartbeat=20)
+NODE = NodeDefinition(os.environ.get('AMQP_HOST', 'rabbitmq'), 'guest', 'guest', heartbeat=20)
 logging.basicConfig(level=logging.DEBUG)